GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - Add quit application (refs/gitbutler/add-quit-application)
   - src/ui/components/command_list.rs
 - Add debug print statement (refs/gitbutler/add-debug-print-statement)
   - src/ui/components/mod.rs

Your previous branch was: refs/heads/develop

The sha for that commit was: 52211f0abe2d10989c5308808ba494e7768d3970

For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch